feat: Add getUserId function to retrieve user ID from Supabase authentication cookies refactor: Update import path for getUserId in application tests
